requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
numpy>=2.0.0
opencv-python>=4.8.1.78
Pillow>=10.0.1
pytesseract>=0.3.10
easyocr>=1.7.1
fasttext>=0.9.2
langdetect>=1.0.9
yolov5>=7.0.13
scipy>=1.11.3
matplotlib>=3.8.0
pandas>=2.1.1
streamlit>=1.28.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.22
aiofiles>=23.2.1
torch>=2.1.0
torchvision>=0.16.0